old-files/llm_semantic_chunker/requirements.txt
# LLM Semantic Chunker Requirements
# Dependencies for LLM-based semantic chunking

# Google Gemini API
google-generativeai>=0.3.0

# Tokenization
tiktoken>=0.5.0

# Scientific computing
numpy>=1.24.0

# PDF processing
pdfplumber>=0.10.0
pdf2image>=1.16.0

# Image processing
Pillow>=10.0.0

# OCR for extracting text from images
easyocr>=1.7.0

# Vector Database
chromadb>=0.4.0

# Environment variables
python-dotenv>=1.0.0
